*: clean rollup configs
- use module instead of jsnext:main
- use cjs format for node libraries
- regroup cjs and es configs
