+++ packages/agent-client-js/package.json
@@ -3,7 +3,7 @@
   "version": "3.0.2",
   "description": "Stratumn Indigo agent client library",
   "main": "lib/stratumn-agent-client.js",
-  "jsnext:main": "lib/stratumn-agent-client.mjs",
+  "module": "lib/stratumn-agent-client.mjs",
   "scripts": {
     "test":
       "mocha --compilers js:babel-core/register --recursive --require babel-polyfill --require should 2> /dev/null",
