+++ packages/agent-client-js/rollup.browser.min.config.js
@@ -1,5 +1,5 @@
-import config from './rollup.browser.config';
 import uglify from 'rollup-plugin-uglify';
+import config from './rollup.browser.config';
 
 config.plugins.push(uglify());
 config.output.file = 'dist/stratumn-agent-client.min.js';
