+++ packages/agent-client-js/rollup.es.config.js
@@ -1,12 +1,13 @@
 import config from './rollup.base.config';
 
 const pkg = require('./package.json');
+
 const external = Object.keys(pkg.dependencies);
 
 config.external = external;
 config.output = {
-  file: pkg['jsnext:main'],
-  format: 'es',
+  file: pkg.module,
+  format: 'es'
 };
 
 export default config;
