+++ packages/agent-client-js/rollup.umd.config.js
@@ -1,6 +1,7 @@
 import config from './rollup.base.config';
 
 const pkg = require('./package.json');
+
 const external = Object.keys(pkg.dependencies);
 
 config.output = {
