+++ packages/angular-mapexplorer/package.json
@@ -6,14 +6,12 @@
   "license": "Apache-2.0",
   "homepage": "",
   "main": "lib/angular-mapexplorer.js",
-  "jsnext:main": "lib/angular-mapexplorer.mjs",
+  "module": "lib/angular-mapexplorer.mjs",
   "scripts": {
     "lint": "eslint src",
     "test": "karma start karma.conf.js",
     "test:ci": "npm run test -- --single-run",
-    "build:npm": "npm run build:npm:es && npm run build:npm:umd",
-    "build:npm:es": "rollup -c rollup.es.config.js",
-    "build:npm:umd": "rollup -c rollup.umd.config.js",
+    "build:npm": "rollup -c rollup.npm.config.js",
     "build:browser": "rollup -c rollup.browser.config.js",
     "build:demo": "rollup -c rollup.browser.demo.config.js",
     "build": "npm run build:npm && npm run build:browser",
