+++ packages/mapexplorer-core/package.json
@@ -4,28 +4,26 @@
   "description": "Core library for building Map Explorer components",
   "main": "lib/mapexplorer-core.js",
   "module": "lib/mapexplorer-core.mjs",
-  "jsnext:main": "lib/mapexplorer-core.mjs",
   "scripts": {
-    "test:unit": "mocha --compilers js:babel-register  --require should test/unit/*",
-    "test:unit:ci": "nyc npm run test:unit && nyc report --reporter=text-lcov > coverage.lcov",
+    "test:unit":
+      "mocha --compilers js:babel-register  --require should test/unit/*",
+    "test:unit:ci":
+      "nyc npm run test:unit && nyc report --reporter=text-lcov > coverage.lcov",
     "test:integration": "karma start karma.conf.js",
     "test:integration:ci": "npm run test:integration -- --single-run",
     "test": "npm run test:unit && npm run test:integration",
     "test:ci": "npm run test:unit:ci && npm run test:integration:ci",
-    "build": "rollup -c rollup.es.config.js && rollup -c rollup.umd.config.js && rollup -c rollup.browser.config.js && rollup -c rollup.browser.min.config.js",
+    "build":
+      "rollup -c rollup.npm.config.js && rollup -c rollup.browser.config.js && rollup -c rollup.browser.min.config.js",
     "prepublish": "npm run build",
-    "postpublish": "aws s3 cp dist/mapexplorer-core.js s3://stratumn-libs/mapexplorer-core.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.min.js s3://stratumn-libs/mapexplorer-core.min.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.js s3://stratumn-libs/mapexplorer-core-${npm_package_version}.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.min.js s3://stratumn-libs/mapexplorer-core-${npm_package_version}.min.js --region 'eu-west-1'"
+    "postpublish":
+      "aws s3 cp dist/mapexplorer-core.js s3://stratumn-libs/mapexplorer-core.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.min.js s3://stratumn-libs/mapexplorer-core.min.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.js s3://stratumn-libs/mapexplorer-core-${npm_package_version}.js --region 'eu-west-1' && aws s3 cp dist/mapexplorer-core.min.js s3://stratumn-libs/mapexplorer-core-${npm_package_version}.min.js --region 'eu-west-1'"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/stratumn/indigo-js.git"
   },
-  "keywords": [
-    "stratumn",
-    "blockchain",
-    "mapexplorer",
-    "chainscript"
-  ],
+  "keywords": ["stratumn", "blockchain", "mapexplorer", "chainscript"],
   "author": "Stratumn Team",
   "license": "Apache-2.0",
   "bugs": {
